// DXC_Frontend/src/api/models/departmentWithOrganizationDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { OrganizationDto } from './organizationDto'

export interface DepartmentWithOrganizationDto {
  publicId?: string
  /** @nullable */
  code: string | null
  /** @nullable */
  name: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  contactEmail?: string | null
  /** @nullable */
  contactPhone?: string | null
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
  organizationPublicId?: string
  /** @nullable */
  organizationName: string | null
  organization?: OrganizationDto
}